fix(spinner): divided the spinner-stories according to their use
